--- src/main/java/com/zerobase/munbanggu/type/AuthType.java
@@ -1,5 +1,0 @@
-package com.zerobase.munbanggu.type;
-
-public enum AuthType {
-    KAKAO,NAVER,GOOGLE,NORMAL
-}
