--- src/main/java/com/zerobase/munbanggu/type/UserType.java
@@ -1,5 +1,0 @@
-package com.zerobase.munbanggu.type;
-
-public enum UserType {
-  USER,ADMIN
-}
