+++ src/main/java/com/zerobase/munbanggu/user/type/UserType.java
@@ -1,0 +1,5 @@
+package com.zerobase.munbanggu.user.type;
+
+public enum UserType {
+  USER,ADMIN
+}
